requirements.txt
python-dotenv>=1.0.0
pydantic>=2.0.0
langchain-core>=0.1.0
langgraph>=0.0.1

langchain-gigachat>=0.0.1
langchain-openai>=0.0.1
langchain-google-genai>=0.0.1
langchain-google-vertexai>=0.0.1

streamlit>=1.28.0

requests>=2.31.0
